backend/requirements.txt
Django==5.0.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
dj-database-url==2.2.0
psycopg[binary]==3.2.3
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.32.3
